SONAR-8857 reduce duplicated code for bulk requests
